requirements.txt
opencv-python
ultralytics
matplotlib
Pillow